+++ internal/controllers/frontend/finance/wallet.go
@@ -11,7 +11,7 @@ import (
 	"finworker/internal/templates"
 )
 
-func (c *controller) Wallet(ctx context.Context, walletId int) (*template.Template, map[string]any, error) {
+func (c *controller) Wallet(ctx context.Context, walletId int64) (*template.Template, map[string]any, error) {
 	c.logger.Debug("frontend.user.controller", zap.String("event", "got request"))
 
 	data := utils.BuildDefaultDataMapFromContext(ctx)
