+++ internal/repositories/wallets/get.go
@@ -6,7 +6,7 @@ import (
 	"finworker/internal/models"
 )
 
-func (repo *Repository) Get(ctx context.Context, id int) (models.Wallet, error) {
+func (repo *Repository) Get(ctx context.Context, id int64) (models.Wallet, error) {
 	var wallet models.Wallet
 	err := repo.db.GetContext(ctx, &wallet, `SELECT * FROM wallets WHERE id = $1`, id)
 	return wallet, err
